@@ -7,28 +7,33 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 import javax.persistence.*;
+import javax.validation.constraints.NotNull;
 import java.time.LocalDateTime;
 
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor
 @Entity
-@Table(name = "t_subscription")
-public class Subscription {
+@NotNull
+@Table(name = "t_cycle_storage")
+public class CycleStorage {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "subscription_no")
+    @Column(name = "cycle_storage_no")
     private Long no;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "user_no")
     private User user;
 
-    @Enumerated(EnumType.STRING)
-    private licence licence;// 구독여부(라이센스 발급) -> ACCESS,DENIED
+    @Column(name = "six_ball")
+    private String sixBall;
 
-    @Column(name = "local_date_time")
-    private LocalDateTime subscriptionDate;
+    @Column(name = "storage_cycle")
+    private String storageCycle;
+
+    @Column(name = "cycle_storage_date")
+    private LocalDateTime storageDate;
 
 }
