@@ -1,11 +1,13 @@
 package lotto.practice.random.repository;
 
 import lotto.practice.random.domain.Ball;
+import lotto.practice.random.entity.Lotto;
 import lotto.practice.random.vo.BallVo;
+import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LottoRepository {
+public interface LottoRepository extends JpaRepository<Lotto, Long> {
 
-    //데이터 저장소
-    Ball numSix(BallVo ballVo);
+
+
 
 }
